Remove unsafe
