--- internal/dic/content_unsafe.go
@@ -1,32 +1,0 @@
-// Copyright 2015 ikawaha
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// 	You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-// +build !appengine
-
-package dic
-
-import (
-	"strings"
-	"unsafe"
-)
-
-func newContents(b []byte) [][]string {
-	str := *(*string)(unsafe.Pointer(&b))
-	rows := strings.Split(str, rowDelimiter)
-	m := make([][]string, len(rows))
-	for i, r := range rows {
-		m[i] = strings.Split(r, colDelimiter)
-	}
-	return m
-}
